Overhaul
I realized there were some significant performance bottlenecks with the approach
I was taking in which `IntervalTree` was an associative type that associated
data with unique intervals.

In this version, `IntervalTree` is simply an indexed collection of interval
types. An interval is any type that derives `AbstractInterval` and implements a
`first` and `last` function. This can still be used an an associative
container, but duplicate keys can now be stored. This should greatly simplify
the implementation of `IntervalCollection` in Bio.

Other changes:
  * Use `Nullable` instead of `Nothing`.
  * Optimizations to insertion make it about 3x faster.