Behavior change of <= in 0.7 is fixed in this
v0.6 implementation:
<=(x, y) = !(y < x)
==(x, y) = x === y

v0.7 implementation:
<=(x, y) = (x < y) | (x == y)
==(x, y) = x === y

Since, isless is defined in the super abstract type and === is computed in the
concrete type there is a mismatch when concrete types are used anywhere for
comparison with the 0.7 fallbacks.

The code has been modified such that the fallbacks are never called in an
incompatible manner. However, the real fix requires all the operators be
sensibly defined for all interval types properly.