Changes pertaining to supporting 0.7 (#37)
* fix README.md

* Fix a bug in firstintersection that can lead to missed intersections.

* Fix a test

* Inline a couple Base.next methods.

* Addressing deprecation of Nullable in 0.7.

* 0.7 syntax fix.

* proceeding to find further.

* trying another fix for 0.7

* Revert "proceeding to find further."

This reverts commit 182718d613155a85c70c4bc0cb0c639c5419bc3d.

* Behavior change of <= in 0.7 is fixed in this

v0.6 implementation:
<=(x, y) = !(y < x)
==(x, y) = x === y

v0.7 implementation:
<=(x, y) = (x < y) | (x == y)
==(x, y) = x === y

Since, isless is defined in the super abstract type and === is computed in the
concrete type there is a mismatch when concrete types are used anywhere for
comparison with the 0.7 fallbacks.

The code has been modified such that the fallbacks are never called in an
incompatible manner. However, the real fix requires all the operators be
sensibly defined for all interval types properly.

* Performance optimization for 0.7.

* Compat.jl compatible changes.

* Adding Compat to REQUIRE

* Compat fixes for 0.6

* Changes as suggested by @ScottPJones wrt introducing back the fast
bitwise copying with unsafe_copyto!.

The performance impact was not found to be substantial using the new
methods. Added @compat for 6.2 compatibility.

* Removing unnecessary @compat.

* Adding the benchmark for performance.

* Adding `notnothing` for efficiency

* Fixing build errors.